Refactor: Split presets registry & UI panel
Those modules where too large to be in one file. Changes included
splitting them into multiple files.
No code changes included.
