+++ crates/app/src/host/ui/registry/presets/mod.rs
@@ -1,0 +1,8 @@
+//! Host registry storage for named preset snapshots.
+
+mod capture;
+mod catalog;
+mod model;
+
+pub use catalog::{PresetRegistry, PresetUpdateOutcome};
+pub use model::{Preset, PresetFilterEntry, PresetSearchValueEntry};
